+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     - mochapack: 2.1.2 to 2.1.4
 - The version of node was updated, must be >= 16.13.2 < 17.
   The version of npm  was also updated, must be >= 8.1.2 < 9. (The node and npm versions are still mandatory via .npmrc.)
+- Migrated the module 3D TerrainLayer from Backbone to Vue. The TerrainLayer uses the masterportalAPI's terrain layer on creation.
 
 ### Deprecated
 
@@ -89,8 +90,6 @@
 
 ### Deprecated
 
-### Deprecated
--
 ### Removed
 - The following NPM packages are removed:
     - mocha-loader